# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RentKaro PG Accommodation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  tenant: "tenant",
  owner: "owner",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  phone?: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  tenant: "tenant",
  owner: "owner",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type PropertyGenderPreference =
  (typeof PropertyGenderPreference)[keyof typeof PropertyGenderPreference];

export const PropertyGenderPreference = {
  male: "male",
  female: "female",
  any: "any",
} as const;

export type PropertyRoomType =
  (typeof PropertyRoomType)[keyof typeof PropertyRoomType];

export const PropertyRoomType = {
  single: "single",
  double: "double",
  triple: "triple",
  shared: "shared",
} as const;

export interface Property {
  id: number;
  title: string;
  description: string;
  city: string;
  address: string;
  rent: number;
  deposit: number;
  genderPreference: PropertyGenderPreference;
  roomType: PropertyRoomType;
  amenities: string[];
  images: string[];
  availability: boolean;
  rating?: number | null;
  reviewCount: number;
  sentimentScore?: number | null;
  ownerId: number;
  ownerName: string;
  ownerPhone?: string | null;
  createdAt: string;
}

export interface Review {
  id: number;
  propertyId: number;
  tenantId: number;
  tenantName: string;
  rating: number;
  comment: string;
  sentimentScore?: number | null;
  sentimentLabel?: string | null;
  createdAt: string;
}

export type PropertyDetail = Property & {
  reviews: Review[];
};

export interface PropertiesListResponse {
  properties: Property[];
  total: number;
  page: number;
  totalPages: number;
}

export type CreatePropertyBodyGenderPreference =
  (typeof CreatePropertyBodyGenderPreference)[keyof typeof CreatePropertyBodyGenderPreference];

export const CreatePropertyBodyGenderPreference = {
  male: "male",
  female: "female",
  any: "any",
} as const;

export type CreatePropertyBodyRoomType =
  (typeof CreatePropertyBodyRoomType)[keyof typeof CreatePropertyBodyRoomType];

export const CreatePropertyBodyRoomType = {
  single: "single",
  double: "double",
  triple: "triple",
  shared: "shared",
} as const;

export interface CreatePropertyBody {
  title: string;
  description: string;
  city: string;
  address: string;
  rent: number;
  deposit: number;
  genderPreference: CreatePropertyBodyGenderPreference;
  roomType: CreatePropertyBodyRoomType;
  amenities?: string[];
  images?: string[];
  availability?: boolean;
}

export type UpdatePropertyBodyGenderPreference =
  (typeof UpdatePropertyBodyGenderPreference)[keyof typeof UpdatePropertyBodyGenderPreference];

export const UpdatePropertyBodyGenderPreference = {
  male: "male",
  female: "female",
  any: "any",
} as const;

export type UpdatePropertyBodyRoomType =
  (typeof UpdatePropertyBodyRoomType)[keyof typeof UpdatePropertyBodyRoomType];

export const UpdatePropertyBodyRoomType = {
  single: "single",
  double: "double",
  triple: "triple",
  shared: "shared",
} as const;

export interface UpdatePropertyBody {
  title?: string;
  description?: string;
  city?: string;
  address?: string;
  rent?: number;
  deposit?: number;
  genderPreference?: UpdatePropertyBodyGenderPreference;
  roomType?: UpdatePropertyBodyRoomType;
  amenities?: string[];
  images?: string[];
  availability?: boolean;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  propertyId: number;
  propertyTitle: string;
  propertyCity: string;
  tenantId: number;
  tenantName: string;
  tenantPhone?: string | null;
  tenantEmail: string;
  message?: string | null;
  status: BookingStatus;
  createdAt: string;
}

export interface CreateBookingBody {
  propertyId: number;
  message?: string;
}

export type UpdateBookingStatusBodyStatus =
  (typeof UpdateBookingStatusBodyStatus)[keyof typeof UpdateBookingStatusBodyStatus];

export const UpdateBookingStatusBodyStatus = {
  approved: "approved",
  rejected: "rejected",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusBody {
  status: UpdateBookingStatusBodyStatus;
}

export interface CreateReviewBody {
  rating: number;
  comment: string;
}

export interface CityCount {
  city: string;
  count: number;
}

export interface OwnerStats {
  totalProperties: number;
  totalInquiries: number;
  pendingInquiries: number;
  approvedBookings: number;
  totalViews: number;
  averageRating?: number | null;
  recentInquiries?: Booking[];
}

export interface PlatformStats {
  totalProperties: number;
  totalCities: number;
  totalTenants: number;
  totalOwners: number;
  totalBookings: number;
}

export type ListPropertiesParams = {
  city?: string;
  minRent?: number;
  maxRent?: number;
  genderPreference?: ListPropertiesGenderPreference;
  amenities?: string;
  availability?: boolean;
  sortBy?: ListPropertiesSortBy;
  page?: number;
  limit?: number;
};

export type ListPropertiesGenderPreference =
  (typeof ListPropertiesGenderPreference)[keyof typeof ListPropertiesGenderPreference];

export const ListPropertiesGenderPreference = {
  male: "male",
  female: "female",
  any: "any",
} as const;

export type ListPropertiesSortBy =
  (typeof ListPropertiesSortBy)[keyof typeof ListPropertiesSortBy];

export const ListPropertiesSortBy = {
  rent_asc: "rent_asc",
  rent_desc: "rent_desc",
  rating: "rating",
  newest: "newest",
} as const;

export type GetRecommendedPropertiesParams = {
  limit?: number;
};